fix issue in retention management and refactor
Summary:
right now if we delete and recreate the table, the seq number of segments got reset to start from 0.
this cause issue in newer segment are considered as older than the last segment we have seen.
this diff would solve such issue

we also refactor a bunch of stuff to base class so the kc retention
manager and server retention manager are mostly using the same logics

Reviewers: tingchen, csoman, bzzhang

Reviewed By: bzzhang

Differential Revision: https://code.uberinternal.com/D3528221