change offset mapping strategy for immutable seg to reduce memory
Summary:
update offset mapping to docId to use array to reduce memory
right now we use hashmap for such mapping. It is very expensive to do so
changing to array for mapping will use 4 byte/document.

We only do this for immutable segment as it is easier to controll their size,
also immutable segment is the majority where we incur the memory cost

Reviewers: csoman, tingchen, bzzhang

Reviewed By: bzzhang

Differential Revision: https://code.uberinternal.com/D3534691